+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# vTBD
+- [BP-1518](https://movai.atlassian.net/browse/BP-1518): Create alerts history API
+  - Introduce BaseQuery and AlertQuery
+
 # v3.6.0
 - [BP-1521](https://movai.atlassian.net/browse/BP-1521): Logic to activate and deactivate alert
 - [BP-1520](https://movai.atlassian.net/browse/BP-1520): Deactivate alerts on start / stop flow